webapp/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
requests==2.21.0
urllib3==1.24.2
Werkzeug==0.15.2
PyYAML==5.1